# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Закрытый API для системы управления перевозками",
        "description": "API documentation for My Application",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "andrew.izmaylov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://transportation.test/public/api/v1/",
            "description": "Local development server"
        },
        {
            "url": "https://transportation.test/public/api/v1/",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/shipper/{transportation_id}/update-transportation": {
            "patch": {
                "tags": [
                    "2. Shipper - Перевозки"
                ],
                "summary": "Изменение заявки на перевозку",
                "description": "Изменяет заявку на перевозку. Идентификатор перевозки передается в параметрах URI",
                "operationId": "fb08c39922be603c57ce88254e49c0f2",
                "parameters": [
                    {
                        "name": "transportation_id",
                        "in": "path",
                        "description": "Transportation ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления объекта перевозки",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Москва - Дубай, лекарства"
                                    },
                                    "pickupFrom": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-01-14 12:00:00"
                                    },
                                    "pickupTo": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-02-14 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transportation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransportationCard"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "TransportationCard": {
                "required": [
                    "id",
                    "type",
                    "attributes"
                ],
                "properties": {
                    "id": {
                        "description": "Transportation ID",
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "example": "Transportation"
                    },
                    "attributes": {
                        "required": [
                            "id",
                            "name",
                            "clientId",
                            "pickupFrom",
                            "pickupTo",
                            "transportationStatus"
                        ],
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "clientId": {
                                "type": "string"
                            },
                            "pickupFrom": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-01 10:00:00"
                            },
                            "pickupTo": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-01 18:00:00"
                            },
                            "transportationStatus": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}